HBase:Fix findbugs warning
Reviewers: #hbase, zhangguanghao1

Reviewed By: #hbase, zhangguanghao1

Differential Revision: https://phabricator.d.xiaomi.net/D200712

+++ hbase-mapreduce/src/main/java/org/apache/hadoop/hbase/mapreduce/fds/FDSRecordReader.java
@@ -70,7 +70,7 @@ public class FDSRecordReader extends RecordReader<ImmutableBytesWritable, Result
       return 0;
     }
     int index = scanner.getFileListIterator().previousIndex();
-    return  (index == -1 ? 0 : index) / fileCount;
+    return  (index == -1 ? 0 : index) / (float)fileCount;
   }
 
   @Override
